feat(theme): adapt to mobile, fix primary color doesn't change
