+++ apps/main/src/blocks/theme-editor/index.ts
@@ -1,0 +1,1 @@
+export * from './theme-editor';
